added state

+++ photos.py
@@ -4,8 +4,9 @@ import os
 import requests
 import time
 import subprocess
-import getpass
-import pwd
+import json
+import signal
+from threading import Event
 
 
 
@@ -23,7 +24,8 @@ service = google_workspace.service.GoogleService('photoslibrary', scopes= [scope
 
 
 
-def main():
+def main(last_photo: str = None):
+    stopper = GracefulStop()
     next_page_token = None
     most_recent_photo = []
     while True:
@@ -33,37 +35,74 @@ def main():
 
         photos = album_data.get('mediaItems', [])
         for photo in photos:
-            filename = photo.get('filename')
-            file_path = f'{WALLPAPER_DIR}/{filename}'
+            photo_id = photo.get('id')
+            if not last_photo:
+                file_path = f'{WALLPAPER_DIR}/{photo_id}'
 
-            # if file is not on disk, download
-            if not os.path.exists(file_path):
-                download_photo(filename, photo.get('baseUrl'))
-            
-            # set as wallpaper
-            subprocess.run(['gsettings', 'set', 'org.gnome.desktop.background', 'picture-uri', f'file:///{file_path}'])
+                # if file is not on disk, download
+                if not os.path.exists(file_path):
+                    download_photo(photo_id, photo.get('baseUrl'))
+                
+                # set as wallpaper
+                subprocess.run(['gsettings', 'set', 'org.gnome.desktop.background', 'picture-uri', f'file:///{file_path}'])
 
-            # delete old photo if not save photos
-            if not SAVE_PHOTOS:
-                most_recent_photo.append(file_path)
-                if len(most_recent_photo) > 1:
-                    os.remove(most_recent_photo.pop(0))
+                # delete old photo if not save photos
+                if not SAVE_PHOTOS:
+                    most_recent_photo.append(file_path)
+                    if len(most_recent_photo) > 1:
+                        os.remove(most_recent_photo.pop(0))
+
+
+                stopper.exit.wait(SLEEP_MINS * 60)
+
+                # Handle Stop
+                if stopper.exit.is_set():
+                    break
+
+            else:
+                if last_photo == photo_id:
+                    last_photo = None
+                    most_recent_photo.append(f'{WALLPAPER_DIR}/{photo_id}')
             
-            time.sleep(SLEEP_MINS * 60)
             
 
         next_page_token = album_data.get('nextPageToken')
 
+        # Handle Stop
+        if stopper.exit.is_set():
+            with open('state.json', 'w') as f:
+                json.dump({'last_photo': photo_id}, f)
+            break
 
 
-def download_photo(filename: str, url: str):
-    data = requests.get(f'{url}=w1000').content
-    with open(f'{WALLPAPER_DIR}/{filename}', "wb") as f:
+def download_photo(photo_id: str, url: str):
+    data = requests.get(f'{url}=w1500-h1000').content
+    with open(f'{WALLPAPER_DIR}/{photo_id}', "wb") as f:
         f.write(data)
+
+
+
+class GracefulStop:
+
+
+    def __init__(self):
+        self.exit = Event()
+        signal.signal(signal.SIGINT, self.quit)
+        signal.signal(signal.SIGTERM, self.quit)
+
+    def quit(self, signum, frame):
+        self.exit.set()
+
 
 
 
 
 
 if __name__ == "__main__":
-    main()
+    with open('state.json', 'r') as f:
+        state_data = json.load(f)
+    if state_data:
+        last_photo = state_data['last_photo']
+    else:
+        last_photo = None
+    main(last_photo= last_photo)
